feat: share Zod schemas between frontend and API
Extract all Zod enum definitions and validation schemas into
validators/shared/ as the single source of truth. Existing
validator files now re-export from shared for backward compatibility.
Frontend forms can import the same schemas used by API route handlers.

Closes #396
